refactor: remove timer color name wrapper
