--- apps/game-client/src/features/timers/constants/color-names.ts
@@ -1,1 +1,0 @@
-export { getDefaultColorName } from "@/features/timers/utils/get-default-color-name";
